+++ .editorconfig
@@ -1,0 +1,29 @@
+# This file is for unifying the coding style for different editors and IDEs
+# editorconfig.org
+
+# WordPress Coding Standards
+# https://make.wordpress.org/core/handbook/coding-standards/
+
+# Based from https://github.com/ntwb/wp.editorconfig/blob/master/.editorconfig
+#
+# Need an IDE plugin? http://editorconfig.org/#download
+
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+indent_style = tab
+
+[*.json]
+indent_style = space
+indent_size = 2
+
+[*.yml]
+indent_style = space
+indent_size = 4
+
+[*.txt,wp-config-sample.php]
+end_of_line = crlf
