+++ includes/scripts.php
@@ -9,7 +9,9 @@
 
 
 // Exit if accessed directly
-if( ! defined( 'ABSPATH' ) ) exit;
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
 
 
 /**
@@ -20,12 +22,12 @@ if( ! defined( 'ABSPATH' ) ) exit;
  * @return      void
  */
 function edd_cr_admin_scripts() {
-    global $post;
+	global $post;
 
-    // Only enqueue if this is the add/edit post/page screen
-    if( is_object( $post ) && isset( $post->ID ) ) {
-        wp_enqueue_script( 'edd-cr', EDD_CONTENT_RESTRICTION_URL . 'assets/js/admin.js', array( 'jquery' ), EDD_CONTENT_RESTRICTION_VER );
-        wp_enqueue_style( 'edd-cr', EDD_CONTENT_RESTRICTION_URL . 'assets/css/admin.css' );
-    }
+	// Only enqueue if this is the add/edit post/page screen
+	if ( is_object( $post ) && isset( $post->ID ) ) {
+		wp_enqueue_script( 'edd-cr', EDD_CONTENT_RESTRICTION_URL . 'assets/js/admin.js', array( 'jquery' ), EDD_CONTENT_RESTRICTION_VER );
+		wp_enqueue_style( 'edd-cr', EDD_CONTENT_RESTRICTION_URL . 'assets/css/admin.css' );
+	}
 }
 add_action( 'admin_enqueue_scripts', 'edd_cr_admin_scripts' );
